+++ src/main/java/com/buildoc/buildocDemo/repositories/RolRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface RolRepository extends JpaRepository<Rol, Integer> {
+public interface RolRepository extends JpaRepository<Rol, Long> {
     Optional<Rol>findByNombre(String nombre);
 
 }
